+++ src/parser/block/heading.test.ts
@@ -48,9 +48,9 @@ describe('Unit: parser/block/heading', () => {
       assert.deepStrictEqual(inspect(parser('# !http://host')), [['<h1 id="index:!http://host">!http://host</h1>'], '']);
       assert.deepStrictEqual(inspect(parser('## !http://host')), [['<h2 id="index:!http://host">!<a href="http://host" target="_blank">http://host</a></h2>'], '']);
       assert.deepStrictEqual(inspect(parser('# a((b))')), [['<h1 id="index:a((b))">a((b))</h1>'], '']);
-      assert.deepStrictEqual(inspect(parser('## a((b))')), [['<h2 id="index:a">a<sup class="annotation">b</sup></h2>'], '']);
+      assert.deepStrictEqual(inspect(parser('## a((b))')), [['<h2 id="index:a((b))">a((b))</h2>'], '']);
       assert.deepStrictEqual(inspect(parser('# a[[b]]')), [['<h1 id="index:a[[b]]">a[[b]]</h1>'], '']);
-      assert.deepStrictEqual(inspect(parser('## a[[b]]')), [['<h2 id="index:a">a<sup class="reference">b</sup></h2>'], '']);
+      assert.deepStrictEqual(inspect(parser('## a[[b]]')), [['<h2 id="index:a[[b]]">a[[b]]</h2>'], '']);
       assert.deepStrictEqual(inspect(parser('###### a')), [['<h6 id="index:a">a</h6>'], '']);
       assert.deepStrictEqual(inspect(parser('# a\n##')), [['<h1 id="index:a">a</h1>', '##'], '']);
       assert.deepStrictEqual(inspect(parser('# a\n## ')), [['<h1 id="index:a">a</h1>', '## '], '']);
@@ -81,7 +81,7 @@ describe('Unit: parser/block/heading', () => {
       assert.deepStrictEqual(inspect(parser('# A [#a((b))]')), [['<h1 id="index:a((b))">A<span class="indexer" data-index="a((b))"></span></h1>'], '']);
       assert.deepStrictEqual(inspect(parser('# A [#a[[b]]]')), [['<h1 id="index:a[[b]]">A<span class="indexer" data-index="a[[b]]"></span></h1>'], '']);
       assert.deepStrictEqual(inspect(parser('# A [#b|#c]')), [['<h1 id="index:c">A<span class="indexer" data-index="c"></span></h1>'], '']);
-      assert.deepStrictEqual(inspect(parser('## a [#b] [#c]')), [['<h2 id="index:c">a <a class="index" href="#index:b">b</a><span class="indexer" data-index="c"></span></h2>'], '']);
+      assert.deepStrictEqual(inspect(parser('## a [#b] [#c]')), [['<h2 id="index:c">a [<a class="hashtag" href="/hashtags/b">#b</a>]<span class="indexer" data-index="c"></span></h2>'], '']);
     });
 
   });